Bugfixes related to size and numbers
+ Added conversion for syncsafe numbers
* Fixed size function, I misunderstood how file sizes were calculated
* Improved convert_bytes function
* Changed types for most variables holding numbers (like position in a
file or number of bytes to read, tag size, and so on) to prevent loss of
information because of too small variable types

+++ include/song.hpp
@@ -43,12 +43,12 @@ public:
     std::string m_track_number = "";
 
     // TODO consider making these strings as well
-    std::uint16_t m_audio_start = 0;
-    std::uint16_t m_duration = 0;
-    std::uint16_t m_delay = 0;
+    std::uint32_t m_audio_start = 0;
+    std::uint32_t m_duration = 0;
+    std::uint32_t m_delay = 0;
 
     // offset of the player counter
-    std::uint16_t m_counter_offset = 0;
+    std::uint32_t m_counter_offset = 0;
 
     std::string m_path = "";
